refactor: split and eliminated redundancy in test cases
